restructure code files
