+++ pkg/hintrunner/hinter/scope.go
@@ -2,35 +2,7 @@ package hinter
 
 import (
 	"fmt"
-
-	mem "github.com/NethermindEth/cairo-vm-go/pkg/vm/memory"
 )
-
-// Global context to keep track of different results across different
-// hints execution.
-type HintRunnerContext struct {
-	DictionaryManager         DictionaryManager
-	SquashedDictionaryManager SquashedDictionaryManager
-	ScopeManager              ScopeManager
-	// points towards free memory of a segment
-	ConstantSizeSegment mem.MemoryAddress
-}
-
-func InitializeDefaultContext() *HintRunnerContext {
-	return &HintRunnerContext{
-		DictionaryManager:         DictionaryManager{},
-		SquashedDictionaryManager: SquashedDictionaryManager{},
-		ScopeManager:              *DefaultNewScopeManager(),
-		ConstantSizeSegment:       mem.UnknownAddress,
-	}
-}
-
-func SetContextWithScope(scope map[string]any) *HintRunnerContext {
-	ctx := HintRunnerContext{
-		ScopeManager: *NewScopeManager(scope),
-	}
-	return &ctx
-}
 
 // ScopeManager handles all operations regarding scopes:
 // - Creating a new scope
